+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 2.0.14
 
-  * Mise en place de l'aide sur les analyses de films
+  * Mise en place de l'initialisation de l'analyse de film.
+  * Mise en place de l'aide sur les analyses de films.
   * Gestion de la validation ou du refus de la candidature aux analyses de films.
 
 ### 2.0.13